added docker file
Dockerized the application, needed to change the host because there are conflicts accessing the localhost from docker apps. Tried to install flask-migrate to make the db migrations easy, but there is a conflict running flask run vs python3 app.py. added some resources with diagrams of approach, and also added a readme